rename: CategoryChallenge, Challenge 폴더 변경
